requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
tensorboard>=2.9.0
tqdm>=4.64.0
PyYAML>=6.0

# Mamba-specific (optional, only needed for Mamba backbones)
# mamba-ssm==2.3.0
# mambavision==1.2.0
# transformers>=4.30.0
